# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Field Inspect API
  version: 0.1.0
  description: Field Inspect offline inspection API. Local loopback access by default;
    configured credentials required for network access. Processing completion is not
    field acceptance. No vehicle control or online package-upload endpoint. New functions
    have not been tested.
paths:
...
  /api/access/audit:
    get:
      summary: Audit Log
      operationId: audit_log_api_access_audit_get
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    OperatorSession:
      type: apiKey
      in: cookie
      name: astra_session
servers:
- url: /
security:
- BearerAuth: []
- OperatorSession: []